+++ flowjob-broker-core/src/main/java/org/limbo/flowjob/broker/core/worker/dispatch/DispatchOption.java
@@ -53,12 +53,12 @@ public class DispatchOption implements Serializable {
     /**
      * 所需的CPU核心数，小于等于0表示此作业未定义CPU需求。在分发作业时，会根据此方法返回的CPU核心需求数量来检测一个worker是否有能力执行此作业。
      */
-    private BigDecimal cpuRequirement;
+    private Float cpuRequirement;
 
     /**
-     * 所需的内存GB数，小于等于0表示此作业未定义内存需求。在分发作业时，会根据此方法返回的内存需求数量来检测一个worker是否有能力执行此作业。
+     * 所需的内存MB数，小于等于0表示此作业未定义内存需求。在分发作业时，会根据此方法返回的内存需求数量来检测一个worker是否有能力执行此作业。
      */
-    private BigDecimal ramRequirement;
+    private Long ramRequirement;
 
     /**
      * tag 过滤器配置
